refactor: make theme folder for related code, address review comments
