+++ dev-client/src/theme/styleKeyLists.ts
@@ -1,0 +1,140 @@
+import {TextStyle, ViewStyle} from 'react-native';
+import {ElementOf} from 'terraso-mobile-client/theme/themeManipulations';
+
+export type ViewSpaceKey = ElementOf<typeof viewSpaceKeys>;
+export const viewSpaceKeys = new Set([
+  'bottom',
+  'end',
+  'flexBasis',
+  'rowGap',
+  'gap',
+  'columnGap',
+  'height',
+  'left',
+  'margin',
+  'marginBottom',
+  'marginEnd',
+  'marginHorizontal',
+  'marginLeft',
+  'marginRight',
+  'marginStart',
+  'marginTop',
+  'marginVertical',
+  'maxHeight',
+  'maxWidth',
+  'minHeight',
+  'minWidth',
+  'padding',
+  'paddingBottom',
+  'paddingEnd',
+  'paddingHorizontal',
+  'paddingLeft',
+  'paddingRight',
+  'paddingStart',
+  'paddingTop',
+  'paddingVertical',
+  'right',
+  'start',
+  'top',
+  'width',
+] as const satisfies readonly (keyof ViewStyle)[]);
+
+export type ViewColorKey = ElementOf<typeof viewColorKeys>;
+export const viewColorKeys = new Set([
+  'backgroundColor',
+  'borderBlockColor',
+  'borderBlockEndColor',
+  'borderBlockStartColor',
+  'borderBottomColor',
+  'borderColor',
+  'borderEndColor',
+  'borderLeftColor',
+  'borderRightColor',
+  'borderStartColor',
+  'borderTopColor',
+] as const satisfies readonly (keyof ViewStyle)[]);
+
+export const viewStyleKeys = new Set([
+  'alignContent',
+  'alignItems',
+  'alignSelf',
+  'aspectRatio',
+  'borderBottomWidth',
+  'borderEndWidth',
+  'borderLeftWidth',
+  'borderRightWidth',
+  'borderStartWidth',
+  'borderTopWidth',
+  'borderWidth',
+  'display',
+  'flex',
+  'flexDirection',
+  'flexGrow',
+  'flexShrink',
+  'flexWrap',
+  'justifyContent',
+  'overflow',
+  'position',
+  'zIndex',
+  'direction',
+  'backfaceVisibility',
+  'borderBottomEndRadius',
+  'borderBottomLeftRadius',
+  'borderBottomRightRadius',
+  'borderBottomStartRadius',
+  'borderCurve',
+  'borderEndEndRadius',
+  'borderEndStartRadius',
+  'borderRadius',
+  'borderStartEndRadius',
+  'borderStartStartRadius',
+  'borderStyle',
+  'borderTopEndRadius',
+  'borderTopLeftRadius',
+  'borderTopRightRadius',
+  'borderTopStartRadius',
+  'opacity',
+  'elevation',
+  'pointerEvents',
+  'transform',
+  'transformOrigin',
+  'transformMatrix',
+  'rotation',
+  'scaleX',
+  'scaleY',
+  'translateX',
+  'translateY',
+  'shadowColor',
+  'shadowOffset',
+  'shadowOpacity',
+  'shadowRadius',
+] as const satisfies readonly Exclude<
+  keyof ViewStyle,
+  ViewColorKey | ViewSpaceKey
+>[]);
+
+export const textStyleKeys = new Set([
+  'color',
+  'fontFamily',
+  'fontSize',
+  'fontStyle',
+  'fontWeight',
+  'letterSpacing',
+  'lineHeight',
+  'textAlign',
+  'textDecorationLine',
+  'textDecorationStyle',
+  'textDecorationColor',
+  'textShadowColor',
+  'textShadowOffset',
+  'textShadowRadius',
+  'textTransform',
+  'userSelect',
+  'fontVariant',
+  'textDecorationColor',
+  'textDecorationStyle',
+  'writingDirection',
+  'textAlignVertical',
+  'verticalAlign',
+  'includeFontPadding',
+] as const satisfies (keyof TextStyle)[]);
